backend/api/requirements.txt
fastapi
uvicorn
pydantic
pandas
geopandas
shapely
geopy
fiona
pyproj
python-dotenv